chore(styling): follow review hints
* rename subprocess-navigation -> drilldown
* add JSDoc comments
