+++ lib/Viewer.js
@@ -4,7 +4,7 @@ import CoreModule from './core';
 import TranslateModule from 'diagram-js/lib/i18n/translate';
 import SelectionModule from 'diagram-js/lib/features/selection';
 import OverlaysModule from 'diagram-js/lib/features/overlays';
-import SubprocessNavigationModule from './features/subprocess-navigation';
+import DrilldownModdule from './features/drilldown';
 
 import BaseViewer from './BaseViewer';
 
@@ -68,7 +68,7 @@ Viewer.prototype._modules = [
   TranslateModule,
   SelectionModule,
   OverlaysModule,
-  SubprocessNavigationModule
+  DrilldownModdule
 ];
 
 // default moddle extensions the viewer is composed of
