+++ lib/features/drilldown/DrilldownCentering.js
@@ -1,11 +1,17 @@
 var defaultPosition = { x: 100, y: 50 };
 
+/**
+ * Move collapsed subprocesses into view when drilling down. Zoom and scroll
+ * are saved in a session.
+ *
+ * @param {eventBus} eventBus
+ * @param {canvas} canvas
+ */
 export default function SubprocessCentering(eventBus, canvas) {
   var currentPlane = 'base';
   var positionMap = {};
 
   eventBus.on('plane.set', function(event) {
-
     var currentViewbox = canvas.viewbox();
     positionMap[currentPlane] = {
       x: currentViewbox.x,
