+++ lib/features/drilldown/SubprocessCompatibility.js
@@ -2,8 +2,8 @@
 import { is } from '../../util/ModelUtil';
 
 /**
- * Hooks into `import.render.start` and creates new planes for collapsed
- * subprocesses with elements on the parent diPlane.
+ * Hook into `import.render.start` and create new planes for diagrams with
+ * collapsed subprocesses and all dis on the same plane.
  *
  * @param {eventBus} eventBus
  * @param {moddle} moddle
@@ -60,12 +60,12 @@ SubprocessCompatibility.prototype.fixPlaneDi = function(plane) {
     if (is(parent, 'bpmn:SubProcess') && parent !== plane.bpmnElement) {
 
       // don't change the array while we iterate over it
-      elementsToMove.push({ element: diElement, parent: parent });
+      elementsToMove.push({ diElement: diElement, parent: parent });
     }
   });
 
   elementsToMove.forEach(function(element) {
-    var diElement = element.element;
+    var diElement = element.diElement;
     var parent = element.parent;
 
     // parent is expanded, get nearest collapsed parent
