+++ lib/features/drilldown/SubprocessFakeFlows.js
@@ -1,7 +1,14 @@
 var FLOW_LENGTH = 50;
 var HIGH_RENDER_PRIORITY = 2000;
 
-export default function SubprocessFlows(eventBus, config) {
+/**
+ * Draw representations of flows on expanded shape of subprocesses. The position
+ * is determined relative to the primary element.
+ *
+ * @param {eventBus} eventBus
+ * @param {config.subprocessFlows} config
+ */
+export default function SubprocessFakeFlows(eventBus, config) {
 
   var mutedStrokeColor = (config && config.mutedStrokeColor) || '#dddddd';
 
@@ -86,7 +93,7 @@ export default function SubprocessFlows(eventBus, config) {
 
 }
 
-SubprocessFlows.$inject = [ 'eventBus', 'config.subprocessFlows' ];
+SubprocessFakeFlows.$inject = [ 'eventBus', 'config.subprocessFlows' ];
 
 
 // helpers
