+++ lib/import/BpmnImporter.js
@@ -130,15 +130,17 @@ BpmnImporter.prototype.add = function(semantic, di, parentElement) {
 
   // ROOT ELEMENT
   // handle the special case that we deal with a
-  // invisible root element (process or collaboration)
+  // invisible root element (process, subprocess or collaboration)
   if (is(di, 'bpmndi:BPMNPlane')) {
 
+    // create a new plane with implicit root for subprocesses. The only child
+    // is the expanded subprocess shape, which contains all further children
+    // TODO(marstamm): ideally, the importer would be plugable and we can move this into a module
     if (is(semantic, 'bpmn:SubProcess')) {
       var plane = this._canvas.createPlane(semantic.id, element);
 
       // Get size of expanded element
       var bounds = getPlaneBounds(di);
-
       var primary = this._elementRegistry.get(semantic.id);
       element = this._elementFactory.createShape(elementData(semantic, di, {
         collapsed: false,
